requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2
python-dotenv
google-genai